refactor(sema): resolve natspec after typeck
